refactor:  switch direction of loading drivers
Rather than having the driver define functions and then load the base
code, have the base code load the driver.  This allows us to clearly
define the API that is expected from drivers.

There's a bit of fallout in this that ideally could have been in smaller
commits, but ... we're only a few commits deep anyways.

- Clean up makefile to install where it's told to rather than assume a
local HOME install
- Drop the poorly implemented and undocumented dbname.  This needs to
come back later with a thought-out commit
- Add DBENV_ROOT as an environment variable that can be used for testing
or to override the search path for database drivers